scrapper/db/migrations/versions/0004_make_category_path_nullable.py
from yoyo import step

__depends__ = {'0001_init_schema'}

steps = [
    step(
        """
        ALTER TABLE categories ALTER COLUMN path DROP NOT NULL;
        """,
        """
        ALTER TABLE categories ALTER COLUMN path SET NOT NULL;
        """
    )
] 