requirements.txt
# Scraping
scrapy>=2.11.0
itemadapter>=0.3.0
selenium
webdriver-manager
sentry-sdk
requests

# Database
supabase
psycopg2-binary>=2.9.1
yoyo-migrations>=7.3.2
python-dotenv>=0.19.0
brotli